# Minor performance output enhancement.