@hz-devops-test 2023-10-20
The job `Simulator-pr-builder` of your PR failed. (Hazelcast internal details: [build log](https://s3.console.aws.amazon.com/s3/buckets/j-artifacts/Simulator-pr-builder/720/Simulator-pr-builder-720.zip), [artifacts](https://s3.console.aws.amazon.com/s3/buckets/j-artifacts/Simulator-pr-builder/720/Simulator-pr-builder-720.zip)).
Through arcane magic we have determined that the following fragments from the build log may contain information about the problem.
<details>
<summary>Click to expand the log file</summary>
<pre>
---------ERRORS-----------
--------------------------
[ERROR] Tests run: 10, Failures: 1, Errors: 0, Skipped: 0, Time elapsed: 0.003 s <<< FAILURE! -- in com.hazelcast.simulator.coordinator.PerformanceStatsCollectorTest
--------------------------
[ERROR] com.hazelcast.simulator.coordinator.PerformanceStatsCollectorTest.testFormatPerformanceNumbers_avgLatencyOverMicrosThreshold -- Time elapsed: 0.002 s <<< FAILURE!
--------------------------
[ERROR] Failures: 
--------------------------
[ERROR]   PerformanceStatsCollectorTest.testFormatPerformanceNumbers_avgLatencyOverMicrosThreshold:82
--------------------------
[ERROR] Tests run: 591, Failures: 1, Errors: 0, Skipped: 3
--------------------------
[ERROR] There are test failures.
--------------------------
--------------------------
-------TEST FAILURE-------
--------------------------
[INFO] Results:
[INFO] 
[ERROR] Failures: 
[ERROR]   PerformanceStatsCollectorTest.testFormatPerformanceNumbers_avgLatencyOverMicrosThreshold:82
[INFO] 
[ERROR] Tests run: 591, Failures: 1, Errors: 0, Skipped: 3
[INFO] 

[ERROR] There are test failures.
</pre>
</details>